+++ changelog.md
@@ -3,8 +3,15 @@ All notable changes after version 1.6.16 to this project will be documented in t
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ##[unreleased] Unreleased
+
+##[1.21.7] 2017-07-07
+### Fixed
+- removed excessive bracket in `WPBrowser` template
+
+##[1.21.6] 2017-07-06
 ### Changed
 - switch to `.env` based configuration for tests
+- fix an issue where in some cases (e.g. CLI) `WPLoader` module set to `loadOnly: true` would generate errors
 
 ##[1.21.5] 2017-06-30
 ### Fixed
@@ -580,7 +587,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Reference to ModuleConfigException class in WPLoader class.
 
-[unreleased]: https://github.com/lucatume/wp-browser/compare/1.21.5...HEAD
+[unreleased]: https://github.com/lucatume/wp-browser/compare/1.21.7...HEAD
+[1.21.7]: https://github.com/lucatume/wp-browser/compare/1.21.6...1.21.7
+[1.21.6]: https://github.com/lucatume/wp-browser/compare/1.21.5...1.21.6
 [1.21.5]: https://github.com/lucatume/wp-browser/compare/1.21.4...1.21.5
 [1.21.4]: https://github.com/lucatume/wp-browser/compare/1.21.2...1.21.4
 [1.21.3]: https://github.com/lucatume/wp-browser/compare/1.21.2...1.21.3